+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add user profile wizard [#729](https://github.com/azavea/echo-locator/pull/729) [#737](https://github.com/azavea/echo-locator/pull/737)
 - Add top 10 tour [#730](https://github.com/azavea/echo-locator/pull/730)
 - Add neighborhood detail preview cards to Discover page [#741](https://github.com/azavea/echo-locator/pull/741)
+- Add base Trips Wizard [#740](https://github.com/azavea/echo-locator/pull/740)
 
 ### Changed